# Speedup (x 3-4) cubic spline 2d