Construction of the spline is especially slow, compared to evaluating it.

This PR allows a speedup of ~3-4x for construction, and a minor speedup in evaluating (5%).

10 million constructions:
old: 46s
new: 13s  (3.5x faster)

This also speeds up downstream tools, e.g. PeakPickerHiRes' picking step:
6.52s vs 8.22s (26% speedup)