@cbielow 2015-05-08
mostly the extra vectors b,c,d which were just temporaries and later 
copied to b_,c_,d_
and then the reserve/resize on the target vectors was missing.
Those two things are the main contributors.

for front/back:
I just find x.back() more readable than x[x.size()-1]
Not sure if its faster -- it probably does not matter much.


Am 5/8/2015 um 11:56 AM schrieb Lars Nilse:
> Out of interest, what was the main culprit? front/back?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/OpenMS/OpenMS/pull/1401#issuecomment-100178132>.
>